+++ src/libstd/std.rc
@@ -7,7 +7,7 @@
 #[license = "MIT"];
 #[crate_type = "lib"];
 
-export fs, io, net, run, uv, uvtmp;
+export fs, io, net, run, uv;
 export c_vec, four, tri, util;
 export bitv, deque, fun_treemap, list, map, smallintmap, sort, treemap, ufind;
 export rope;
@@ -25,7 +25,6 @@ mod net;
 #[path =  "run_program.rs"]
 mod run;
 mod uv;
-mod uvtmp;
 
 
 // Utility modules
