+++ src/rt/rustrt.def.in
@@ -80,38 +80,19 @@ upcall_call_shim_on_rust_stack
 upcall_new_stack
 upcall_del_stack
 upcall_reset_stack_limit
-rust_uv_default_loop
 rust_uv_loop_new
-rust_uv_loop_delete
+rust_uv_loop_set_data
+rust_uv_bind_op_cb
+rust_uv_stop_op_cb
 rust_uv_run
-rust_uv_unref
-rust_uv_idle_init
-rust_uv_idle_start
-rust_uv_size_of_idle_t
-rust_uvtmp_create_thread
-rust_uvtmp_start_thread
-rust_uvtmp_join_thread
-rust_uvtmp_delete_thread
-rust_uvtmp_connect
-rust_uvtmp_close_connection
-rust_uvtmp_write
-rust_uvtmp_read_start
-rust_uvtmp_timer
-rust_uvtmp_delete_buf
-rust_uvtmp_get_req_id
-rust_uvtmp_uv_loop_new
-rust_uvtmp_uv_loop_set_data
-rust_uvtmp_uv_bind_op_cb
-rust_uvtmp_uv_stop_op_cb
-rust_uvtmp_uv_run
-rust_uvtmp_uv_close
-rust_uvtmp_uv_close_async
-rust_uvtmp_uv_close_timer
-rust_uvtmp_uv_async_send
-rust_uvtmp_uv_async_init
-rust_uvtmp_uv_timer_init
-rust_uvtmp_uv_timer_start
-rust_uvtmp_uv_timer_stop
+rust_uv_close
+rust_uv_close_async
+rust_uv_close_timer
+rust_uv_async_send
+rust_uv_async_init
+rust_uv_timer_init
+rust_uv_timer_start
+rust_uv_timer_stop
 rust_dbg_lock_create
 rust_dbg_lock_destroy
 rust_dbg_lock_lock
